Feat: InvitationLabel 삭제, InvitationText 컴포넌트 디자인 구현
